+++ kafka_consumer/datadog_checks/kafka_consumer/client/confluent_kafka_client.py
@@ -1,0 +1,21 @@
+# (C) Datadog, Inc. 2023-present
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
+class ConfluentKafkaClient:
+    def __init__(self) -> None:
+        pass
+
+    def get_consumer_offsets(self):
+        pass
+
+    def get_broker_offset(self):
+        pass
+
+    def report_consumer_offset_and_lag(self):
+        pass
+
+    def report_broker_offset(self):
+        pass
+
+    def collect_broker_metadata(self):
+        pass
